Refactor web3auth implementation and clean up chain definitions
