+++ apps/iframe/.env.example
@@ -32,8 +32,6 @@ VITE_WEB3AUTH_VERIFIER=
 #######################################
 VITE_SUBMITTER_URL=https://submitter.happy.tech
 
-VITE_CHAIN_ID=31337
-
 #######################################
 ################ FAUCET ###############
 #######################################
