+++ apps/iframe/src/connections/firebase/workers/firebase/firebaseAuthState.ts
