+++ apps/iframe/src/connections/firebase/workers/web3auth/polyfill.ts
